refactor: useTeams, useTeam 파일 분리
