+++ frontend/src/pages/teams/InterviewTeamAdd.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useEffect } from 'react';
 
-import { useTeam } from 'hooks/useTeams';
+import useTeam from 'hooks/useTeam';
 
 import Button from 'components/@commons/Button';
 import ContentHeader from 'components/@commons/ContentHeader';
